fix: refactoring reducers, change modal and modalOverlay, change quantity bun's, added navlink for Logo
