+++ src/services/actions/index.js
@@ -8,12 +8,6 @@ export const GET_ITEMS_ERROR = 'GET_ITEMS_ERROR';
 export const ORDER_REQUEST = 'ORDER_REQUEST';
 export const ORDER_SUCCESS = 'ORDER_SUCCESS';
 export const ORDER_ERROR = 'ORDER_ERROR';
-export const ADD_INGREDIENT = 'ADD_INGREDIENT';
-export const DELETE_INGREDIENT = 'DELETE_INGREDIENT';
-export const OPEN_MODAL_ORDER = 'OPEN_MODAL_ORDER';
-export const CLOSE_MODAL_ORDER = 'CLOSE_MODAL_ORDER';
-export const OPEN_MODAL_INGREDIENT = 'OPEN_MODAL_INGREDIENT';
-export const CLOSE_MODAL_INGREDIENT = 'CLOSE_MODAL_INGREDIENT';
 export const ADDED_ITEM = 'ADDED_ITEM';
 export const DELETE_ITEM_ON_INDEX = 'DELETE_ITEM_ON_INDEX';
 export const CHANGE_ORDER_INGREDIENT_IN_BURGER = 'CHANGE_ORDER_INGREDIENT_IN_BURGER';
@@ -73,9 +67,6 @@ export function gerOrder(idIngredients) {
                     dispatch({
                         type: ORDER_SUCCESS,
                         orders: res
-                    });
-                    dispatch({
-                        type: OPEN_MODAL_ORDER
                     });
                 } else {
                     dispatch({
