+++ src/services/reducers/items.js
@@ -1,0 +1,34 @@
+import { GET_ITEMS_REQUEST, GET_ITEMS_SUCCESS, GET_ITEMS_ERROR } from '../actions/index';
+const initialStateItems = {
+    items: [],
+    isLoading: false,
+    itemsError: false
+};
+export const itemsReducer = (state = initialStateItems, action) => {
+    switch (action.type) {
+        case GET_ITEMS_REQUEST: {
+            return {
+                ...state,
+                isLoading: false
+            }
+        }
+        case GET_ITEMS_SUCCESS: {
+            return {
+                ...state,
+                isLoading: true,
+                itemsError: false,
+                items: action.items
+            }
+        }
+        case GET_ITEMS_ERROR: {
+            return {
+                ...state,
+                isLoading: false,
+                itemsError: true
+            }
+        }
+        default: {
+            return state;
+        }
+    }
+}
